refactor: refactor related Notification
